+++ bgpd/bgp_mplsvpn.h
@@ -22,15 +22,7 @@
 #define _QUAGGA_BGP_MPLSVPN_H
 
 #include "bgpd/bgp_route.h"
-
-#define RD_TYPE_AS      0
-#define RD_TYPE_IP      1
-#define RD_TYPE_AS4     2
-#if ENABLE_BGP_VNC
-#define RD_TYPE_VNC_ETH	0xff00  /* VNC L2VPN */
-#endif
-
-#define RD_ADDRSTRLEN  28
+#include "bgpd/bgp_rd.h"
 
 #ifdef MPLS_LABEL_MAX
 # undef MPLS_LABEL_MAX
@@ -67,44 +59,10 @@ typedef enum {
   "Address Family\n"   \
   "Address Family\n"
 
-struct rd_as
-{
-  u_int16_t type;
-  as_t as;
-  u_int32_t val;
-};
-
-struct rd_ip
-{
-  u_int16_t type;
-  struct in_addr ip;
-  u_int16_t val;
-};
-
-#if ENABLE_BGP_VNC
-struct rd_vnc_eth
-{
-  u_int16_t type;
-  uint8_t local_nve_id;
-  struct ethaddr macaddr;
-};
-#endif
-
-extern u_int16_t decode_rd_type (u_char *);
-extern void encode_rd_type (u_int16_t, u_char *);
 extern void bgp_mplsvpn_init (void);
 extern int bgp_nlri_parse_vpn (struct peer *, struct attr *, struct bgp_nlri *);
 extern u_int32_t decode_label (u_char *);
 extern void encode_label(u_int32_t, u_char *);
-extern void decode_rd_as (u_char *, struct rd_as *);
-extern void decode_rd_as4 (u_char *, struct rd_as *);
-extern void decode_rd_ip (u_char *, struct rd_ip *);
-#if ENABLE_BGP_VNC
-extern void
-decode_rd_vnc_eth (u_char *pnt, struct rd_vnc_eth *rd_vnc_eth);
-#endif
-extern int str2prefix_rd (const char *, struct prefix_rd *);
-extern char *prefix_rd2str (struct prefix_rd *, char *, size_t);
 
 extern int
 argv_find_and_parse_vpnvx(struct cmd_token **argv, int argc, int *index, afi_t *afi);
